+++ spec/integration/unique_violation_integration_spec.rb
@@ -29,6 +29,7 @@ module App exposing (add)
 import Sql exposing (
   Assignable,
   Assignment,
+  Col(..),
   Expr,
   NoJoins,
   Pk,